rename plugin file
