+++ pytemscript/plugins/tecnai_ccd.py
